Add '+ 0' to POSIX/XOPEN option macros
While POSIX.1-2001 onward requires the POSIX and XOPEN option macros to have
specified values (e.g. _POSIX_FSYNC must be -1, 0, or _POSIX_VERSION),
POSIX.1-1996 gives them 'unspecified' values if defined. So for the possibility
that an implementation defines them as empty, add a '+ 0' to the definitions of
their CZ_* wrapped counterparts.